Move into a public command
